fixBugs&update: 移动交叉熵并修复计算问题 将类移动至 whu.edu.cn.util.Entropy
